refactor(board-application): use base-package as `nettee.board.application`
