+++ services/board/application/src/main/java/nettee/board/application/port/BoardCommandRepositoryPort.java
@@ -1,4 +1,4 @@
-package nettee.board.port;
+package nettee.board.application.port;
 
 import nettee.board.domain.Board;
