+++ services/board/application/src/main/java/nettee/board/application/usecase/BoardCreateUseCase.java
@@ -1,4 +1,4 @@
-package nettee.board.usecase;
+package nettee.board.application.usecase;
 
 import nettee.board.domain.Board;
