+++ services/board/driving/web-mvc/src/main/java/nettee/board/web/BoardCommandApi.java
@@ -3,7 +3,7 @@ package nettee.board.web;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import nettee.board.domain.type.BoardStatus;
-import nettee.board.usecase.BoardCreateUseCase;
+import nettee.board.application.usecase.BoardCreateUseCase;
 import nettee.board.web.dto.BoardCommandDto.BoardCreateCommand;
 import nettee.board.web.dto.BoardCommandDto.BoardCreateResponse;
 import nettee.board.web.mapper.BoardDtoMapper;
